requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
cloudinary==1.44.1
python-dotenv==1.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
requests==2.32.5
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise
django-allauth
dj-rest-auth
cryptography
